pyproject.toml
[project]
name = "rag-ai-agent"
version = "0.1.0"
description = "RAG-based AI agent for task tracking and management"
authors = [
    {name = "amanueldemirew", email = "amanuel.demirew@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "llama-index>=0.9.0",
    "llama-index-embeddings-gemini",
    "llama-index-llms-gemini",
    "llama-index-readers-database",
    "llama-index-vector-stores-postgres",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-ai (>=0.0.29,<0.0.30)",
]

[tool.poetry]
packages = [
    {include = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
